fix(native): mega-pass — recycling leaks, cache collisions, per-keystroke cost
Android editor:
- insertMarkdown (every JS paste) now splices lineBlocks before copying
  decoded blocks; multi-line pastes no longer strip block formatting from
  every line below the caret
- plain keystrokes rebuild derived spans only for the touched lines
  (strict-overlap expansion; structural edits keep the full pass), the
  interval scan is a sorted sweep instead of O(spans²), and the whole-doc
  markdown serialization is coalesced to one JNI round trip per frame
- resetForRecycle runs watcher-silent (no phantom events to the old
  component) and resets every prop-backed field (placeholder, link color,
  input type, mention triggers, cursor/selection tint, scroll state)
- IME mark-restore snapshot capped at 48 chars so clipboard-chip pastes
  over a selection don't get restamped with stale marks; armed-mark guard
  position invalidated by every edit; nested edits can't clobber watcher
  bookkeeping
- scrollEnabled actually gates internal scrolling (was wired to scrollBAR
  visibility and clobbered by every height publish); selection survives
  input-type changes; setStylesJson restyles existing content
- color props reset to platform defaults when cleared; requestLayout posts
  deduplicated; block-decoration drawing early-outs for plain documents

Android viewer: markdown changes clear revealed spoilers and learned image
sizes; recycle resets allowFontScaling/stateWrapper; discarded image views
cancel their Glide requests; detached BlockTextViews cancel pending
long-presses. JNI: attached threads detach via pthread key; measure guards
OOM'd arrays before calling into Java. Dead code dropped (plain-text
bridge, unused span constants).

iOS editor:
- line iteration is terminator-safe: a \r\n reaching the storage (drag &
  drop, JS-provided labels) hung the main thread in five loops; CRs are
  also normalized to \n at the door and stripped from link/mention labels
- links/mentions refuse code-block context like every other mark path;
  markdown emission coalesced per runloop turn; selection events only fire
  when the selection actually moved; pasted images PNG-encode off-main;
  the focus-recycle cycle emits blur if refocusing fails; emoji mention
  triggers keep their full composed character

iOS viewer: content/bound cache keys use full contents (CFString.hash
samples 96 chars — long documents collided and rendered each other), image
size keys use sorted contents (NSDictionary.hash is the entry count), an
images-prop change forces relayout, markdown changes clear revealed
spoilers, spoiler toggles only redraw text views.

Both shadow nodes clamp non-finite width constraints.

Gates: 133/133 C++, 17 JS, tsc; both example apps device-verified.

+++ cpp/react/FastMarkdownShadowNode.cpp
@@ -76,7 +76,7 @@ Size FastMarkdownShadowNode::measureContent(
   // viewport-ish minimum and let Yoga assign the real frame.
   float maxWidth = layoutConstraints.maximumSize.width;
   if (!std::isfinite(maxWidth)) {
-    maxWidth = std::max(layoutConstraints.minimumSize.width, 0.0f);
+    maxWidth = std::max(static_cast<float>(layoutConstraints.minimumSize.width), 0.0f);
   }
   // The host views compute the same multiplier (RN's iOS category table /
   // Android configuration.fontScale) so measured and rendered heights
